requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
python-dotenv~=1.1.0
jinja2
pydantic[email]~=2.11.3
sqlalchemy~=2.0.40
postgres
psycopg2
bcrypt~=4.3.0
cryptography~=44.0.2
PyJWT~=2.10.1
redis
python-multipart